Move `RustGenerator` out of `rust-lib`
