+++ crates/rust/src/interface.rs
@@ -2,14 +2,14 @@ use crate::bindgen::FunctionBindgen;
 use crate::{Direction, ExportKey, Identifier, RustWasm};
 use anyhow::Result;
 use heck::*;
-use std::collections::BTreeMap;
+use std::collections::{BTreeMap, BTreeSet};
 use std::fmt::Write as _;
 use std::mem;
 use wit_bindgen_core::abi::{self, AbiVariant, LiftLower};
-use wit_bindgen_core::{uwrite, uwriteln, wit_parser::*, Source, TypeInfo, Types};
+use wit_bindgen_core::{uwrite, uwriteln, wit_parser::*, Source, TypeInfo};
 use wit_bindgen_rust_lib::{
     dealias, int_repr, to_rust_ident, to_upper_camel_case, wasm_type, FnSig, Ownership,
-    RustFlagsRepr, RustGenerator, TypeMode,
+    RustFlagsRepr, TypeMode,
 };
 
 pub struct InterfaceGenerator<'a> {
@@ -461,20 +461,1029 @@ impl InterfaceGenerator<'_> {
 
         self.src.push_str("}\n");
     }
-}
 
-impl<'a> RustGenerator<'a> for InterfaceGenerator<'a> {
-    fn resolve(&self) -> &'a Resolve {
-        self.resolve
+    fn rustdoc(&mut self, docs: &Docs) {
+        let docs = match &docs.contents {
+            Some(docs) => docs,
+            None => return,
+        };
+        for line in docs.trim().lines() {
+            self.push_str("/// ");
+            self.push_str(line);
+            self.push_str("\n");
+        }
     }
 
-    fn ownership(&self) -> Ownership {
-        self.gen.opts.ownership
+    fn rustdoc_params(&mut self, docs: &[(String, Type)], header: &str) {
+        let _ = (docs, header);
+        // let docs = docs
+        //     .iter()
+        //     .filter(|param| param.docs.trim().len() > 0)
+        //     .collect::<Vec<_>>();
+        // if docs.len() == 0 {
+        //     return;
+        // }
+
+        // self.push_str("///\n");
+        // self.push_str("/// ## ");
+        // self.push_str(header);
+        // self.push_str("\n");
+        // self.push_str("///\n");
+
+        // for param in docs {
+        //     for (i, line) in param.docs.lines().enumerate() {
+        //         self.push_str("/// ");
+        //         // Currently wasi only has at most one return value, so there's no
+        //         // need to indent it or name it.
+        //         if header != "Return" {
+        //             if i == 0 {
+        //                 self.push_str("* `");
+        //                 self.push_str(to_rust_ident(param.name.as_str()));
+        //                 self.push_str("` - ");
+        //             } else {
+        //                 self.push_str("  ");
+        //             }
+        //         }
+        //         self.push_str(line);
+        //         self.push_str("\n");
+        //     }
+        // }
     }
 
-    fn additional_derives(&self) -> &[String] {
-        &self.gen.opts.additional_derive_attributes
+    fn print_signature(
+        &mut self,
+        func: &Function,
+        param_mode: TypeMode,
+        sig: &FnSig,
+    ) -> Vec<String> {
+        let params = self.print_docs_and_params(func, param_mode, sig);
+        if let FunctionKind::Constructor(_) = &func.kind {
+            self.push_str(" -> Self")
+        } else {
+            self.print_results(&func.results, TypeMode::Owned);
+        }
+        params
     }
+
+    fn print_docs_and_params(
+        &mut self,
+        func: &Function,
+        param_mode: TypeMode,
+        sig: &FnSig,
+    ) -> Vec<String> {
+        self.rustdoc(&func.docs);
+        self.rustdoc_params(&func.params, "Parameters");
+        // TODO: re-add this when docs are back
+        // self.rustdoc_params(&func.results, "Return");
+
+        if !sig.private {
+            self.push_str("pub ");
+        }
+        if sig.unsafe_ {
+            self.push_str("unsafe ");
+        }
+        if sig.async_ {
+            self.push_str("async ");
+        }
+        self.push_str("fn ");
+        let func_name = if sig.use_item_name {
+            if let FunctionKind::Constructor(_) = &func.kind {
+                "new"
+            } else {
+                func.item_name()
+            }
+        } else {
+            &func.name
+        };
+        self.push_str(&to_rust_ident(func_name));
+        if let Some(generics) = &sig.generics {
+            self.push_str(generics);
+        }
+        self.push_str("(");
+        if let Some(arg) = &sig.self_arg {
+            self.push_str(arg);
+            self.push_str(",");
+        }
+        let mut params = Vec::new();
+        for (i, (name, param)) in func.params.iter().enumerate() {
+            if i == 0 && sig.self_is_first_param {
+                params.push("self".to_string());
+                continue;
+            }
+            let name = to_rust_ident(name);
+            self.push_str(&name);
+            params.push(name);
+            self.push_str(": ");
+            self.print_ty(param, param_mode);
+            self.push_str(",");
+        }
+        self.push_str(")");
+        params
+    }
+
+    fn print_results(&mut self, results: &Results, mode: TypeMode) {
+        match results.len() {
+            0 => {}
+            1 => {
+                self.push_str(" -> ");
+                self.print_ty(results.iter_types().next().unwrap(), mode);
+            }
+            _ => {
+                self.push_str(" -> (");
+                for ty in results.iter_types() {
+                    self.print_ty(ty, mode);
+                    self.push_str(", ")
+                }
+                self.push_str(")")
+            }
+        }
+    }
+
+    fn print_ty(&mut self, ty: &Type, mode: TypeMode) {
+        match ty {
+            Type::Id(t) => self.print_tyid(*t, mode),
+            Type::Bool => self.push_str("bool"),
+            Type::U8 => self.push_str("u8"),
+            Type::U16 => self.push_str("u16"),
+            Type::U32 => self.push_str("u32"),
+            Type::U64 => self.push_str("u64"),
+            Type::S8 => self.push_str("i8"),
+            Type::S16 => self.push_str("i16"),
+            Type::S32 => self.push_str("i32"),
+            Type::S64 => self.push_str("i64"),
+            Type::Float32 => self.push_str("f32"),
+            Type::Float64 => self.push_str("f64"),
+            Type::Char => self.push_str("char"),
+            Type::String => match mode {
+                TypeMode::AllBorrowed(lt) | TypeMode::LeafBorrowed(lt) => {
+                    self.print_borrowed_str(lt)
+                }
+                TypeMode::Owned | TypeMode::HandlesBorrowed(_) => {
+                    if self.gen.opts.raw_strings {
+                        self.push_vec_name();
+                        self.push_str("::<u8>");
+                    } else {
+                        self.push_string_name();
+                    }
+                }
+            },
+        }
+    }
+
+    fn print_optional_ty(&mut self, ty: Option<&Type>, mode: TypeMode) {
+        match ty {
+            Some(ty) => self.print_ty(ty, mode),
+            None => self.push_str("()"),
+        }
+    }
+
+    pub fn type_path(&self, id: TypeId, owned: bool) -> String {
+        self.type_path_with_name(
+            id,
+            if owned {
+                self.result_name(id)
+            } else {
+                self.param_name(id)
+            },
+        )
+    }
+
+    fn type_path_with_name(&self, id: TypeId, name: String) -> String {
+        if let TypeOwner::Interface(id) = self.resolve.types[id].owner {
+            if let Some(path) = self.path_to_interface(id) {
+                return format!("{path}::{name}");
+            }
+        }
+        name
+    }
+
+    fn print_tyid(&mut self, id: TypeId, mode: TypeMode) {
+        let info = self.info(id);
+        let lt = self.lifetime_for(&info, mode);
+        let ty = &self.resolve.types[id];
+        if ty.name.is_some() {
+            // If `mode` is borrowed then that means literal ownership of the
+            // input type is not necessarily required. In this situation we
+            // ideally want to put a `&` in front to statically indicate this.
+            // That's not required in all situations however and is only really
+            // critical for lists which otherwise would transfer ownership of
+            // the allocation to this function.
+            //
+            // Note, though, that if the type has an `own<T>` inside of it then
+            // it is actually required that we take ownership since Rust is
+            // losing access to those handles.
+            //
+            // Check here if the type has the right shape and if we're in the
+            // right mode, and if those conditions are met a lifetime is
+            // printed.
+            if info.has_list && !info.has_own_handle {
+                if let TypeMode::AllBorrowed(lt)
+                | TypeMode::LeafBorrowed(lt)
+                | TypeMode::HandlesBorrowed(lt) = mode
+                {
+                    self.push_str("&");
+                    if lt != "'_" {
+                        self.push_str(lt);
+                        self.push_str(" ");
+                    }
+                }
+            }
+            let name = self.type_path(id, lt.is_none());
+            self.push_str(&name);
+
+            // If the type recursively owns data and it's a
+            // variant/record/list, then we need to place the
+            // lifetime parameter on the type as well.
+            if (info.has_list || info.has_borrow_handle)
+                && !info.has_own_handle
+                && needs_generics(self.resolve, &ty.kind)
+            {
+                self.print_generics(lt);
+            }
+
+            return;
+
+            fn needs_generics(resolve: &Resolve, ty: &TypeDefKind) -> bool {
+                match ty {
+                    TypeDefKind::Variant(_)
+                    | TypeDefKind::Record(_)
+                    | TypeDefKind::Option(_)
+                    | TypeDefKind::Result(_)
+                    | TypeDefKind::Future(_)
+                    | TypeDefKind::Stream(_)
+                    | TypeDefKind::List(_)
+                    | TypeDefKind::Flags(_)
+                    | TypeDefKind::Enum(_)
+                    | TypeDefKind::Tuple(_) => true,
+                    TypeDefKind::Type(Type::Id(t)) => {
+                        needs_generics(resolve, &resolve.types[*t].kind)
+                    }
+                    TypeDefKind::Type(Type::String) => true,
+                    TypeDefKind::Handle(Handle::Borrow(_)) => true,
+                    TypeDefKind::Resource | TypeDefKind::Handle(_) | TypeDefKind::Type(_) => false,
+                    TypeDefKind::Unknown => unreachable!(),
+                }
+            }
+        }
+
+        match &ty.kind {
+            TypeDefKind::List(t) => self.print_list(t, mode),
+
+            TypeDefKind::Option(t) => {
+                self.push_str("Option<");
+                self.print_ty(t, mode);
+                self.push_str(">");
+            }
+
+            TypeDefKind::Result(r) => {
+                self.push_str("Result<");
+                self.print_optional_ty(r.ok.as_ref(), mode);
+                self.push_str(",");
+                self.print_optional_ty(r.err.as_ref(), mode);
+                self.push_str(">");
+            }
+
+            TypeDefKind::Variant(_) => panic!("unsupported anonymous variant"),
+
+            // Tuple-like records are mapped directly to Rust tuples of
+            // types. Note the trailing comma after each member to
+            // appropriately handle 1-tuples.
+            TypeDefKind::Tuple(t) => {
+                self.push_str("(");
+                for ty in t.types.iter() {
+                    self.print_ty(ty, mode);
+                    self.push_str(",");
+                }
+                self.push_str(")");
+            }
+            TypeDefKind::Resource => {
+                panic!("unsupported anonymous type reference: resource")
+            }
+            TypeDefKind::Record(_) => {
+                panic!("unsupported anonymous type reference: record")
+            }
+            TypeDefKind::Flags(_) => {
+                panic!("unsupported anonymous type reference: flags")
+            }
+            TypeDefKind::Enum(_) => {
+                panic!("unsupported anonymous type reference: enum")
+            }
+            TypeDefKind::Future(ty) => {
+                self.push_str("Future<");
+                self.print_optional_ty(ty.as_ref(), mode);
+                self.push_str(">");
+            }
+            TypeDefKind::Stream(stream) => {
+                self.push_str("Stream<");
+                self.print_optional_ty(stream.element.as_ref(), mode);
+                self.push_str(",");
+                self.print_optional_ty(stream.end.as_ref(), mode);
+                self.push_str(">");
+            }
+
+            TypeDefKind::Handle(Handle::Own(ty)) => {
+                self.mark_resource_owned(*ty);
+                self.print_ty(&Type::Id(*ty), mode);
+            }
+
+            TypeDefKind::Handle(Handle::Borrow(ty)) => {
+                self.push_str("&");
+                if let TypeMode::AllBorrowed(lt)
+                | TypeMode::LeafBorrowed(lt)
+                | TypeMode::HandlesBorrowed(lt) = mode
+                {
+                    if lt != "'_" {
+                        self.push_str(lt);
+                        self.push_str(" ");
+                    }
+                }
+                if self.is_exported_resource(*ty) {
+                    self.push_str(
+                        &self.type_path_with_name(
+                            *ty,
+                            self.resolve.types[*ty]
+                                .name
+                                .as_deref()
+                                .unwrap()
+                                .to_upper_camel_case(),
+                        ),
+                    );
+                } else {
+                    self.print_ty(&Type::Id(*ty), mode);
+                }
+            }
+
+            TypeDefKind::Type(t) => self.print_ty(t, mode),
+
+            TypeDefKind::Unknown => unreachable!(),
+        }
+    }
+
+    fn print_list(&mut self, ty: &Type, mode: TypeMode) {
+        let next_mode = if matches!(self.gen.opts.ownership, Ownership::Owning) {
+            if let TypeMode::HandlesBorrowed(_) = mode {
+                mode
+            } else {
+                TypeMode::Owned
+            }
+        } else {
+            mode
+        };
+        // Lists with `own` handles must always be owned
+        let mode = match *ty {
+            Type::Id(id) if self.info(id).has_own_handle => TypeMode::Owned,
+            _ => mode,
+        };
+        match mode {
+            TypeMode::AllBorrowed(lt) => {
+                self.print_borrowed_slice(false, ty, lt, next_mode);
+            }
+            TypeMode::LeafBorrowed(lt) => {
+                if self.resolve.all_bits_valid(ty) {
+                    self.print_borrowed_slice(false, ty, lt, next_mode);
+                } else {
+                    self.push_vec_name();
+                    self.push_str("::<");
+                    self.print_ty(ty, next_mode);
+                    self.push_str(">");
+                }
+            }
+            TypeMode::Owned | TypeMode::HandlesBorrowed(_) => {
+                self.push_vec_name();
+                self.push_str("::<");
+                self.print_ty(ty, next_mode);
+                self.push_str(">");
+            }
+        }
+    }
+
+    fn print_rust_slice(&mut self, mutbl: bool, ty: &Type, lifetime: &'static str, mode: TypeMode) {
+        self.push_str("&");
+        if lifetime != "'_" {
+            self.push_str(lifetime);
+            self.push_str(" ");
+        }
+        if mutbl {
+            self.push_str(" mut ");
+        }
+        self.push_str("[");
+        self.print_ty(ty, mode);
+        self.push_str("]");
+    }
+
+    fn print_generics(&mut self, lifetime: Option<&str>) {
+        if lifetime.is_none() {
+            return;
+        }
+        self.push_str("<");
+        if let Some(lt) = lifetime {
+            self.push_str(lt);
+            self.push_str(",");
+        }
+        self.push_str(">");
+    }
+
+    fn int_repr(&mut self, repr: Int) {
+        self.push_str(int_repr(repr));
+    }
+
+    fn modes_of(&self, ty: TypeId) -> Vec<(String, TypeMode)> {
+        let info = self.info(ty);
+        // If this type isn't actually used, no need to generate it.
+        if !info.owned && !info.borrowed {
+            return Vec::new();
+        }
+        let mut result = Vec::new();
+
+        // Prioritize generating an "owned" type. This is done to simplify
+        // generated bindings by default. Borrowed handles always use a borrow,
+        // however.
+        let first_mode = if info.owned
+            || !info.borrowed
+            || matches!(self.gen.opts.ownership, Ownership::Owning)
+            || info.has_own_handle
+        {
+            if info.has_borrow_handle {
+                TypeMode::HandlesBorrowed("'a")
+            } else {
+                TypeMode::Owned
+            }
+        } else {
+            assert!(!self.uses_two_names(&info));
+            TypeMode::AllBorrowed("'a")
+        };
+        result.push((self.result_name(ty), first_mode));
+        if self.uses_two_names(&info) {
+            result.push((self.param_name(ty), TypeMode::AllBorrowed("'a")));
+        }
+        result
+    }
+
+    fn print_typedef_record(
+        &mut self,
+        id: TypeId,
+        record: &Record,
+        docs: &Docs,
+        derive_component: bool,
+    ) {
+        let info = self.info(id);
+        // We use a BTree set to make sure we don't have any duplicates and we have a stable order
+        let additional_derives: BTreeSet<String> = self
+            .gen
+            .opts
+            .additional_derive_attributes
+            .iter()
+            .cloned()
+            .collect();
+        for (name, mode) in self.modes_of(id) {
+            let lt = self.lifetime_for(&info, mode);
+            self.rustdoc(docs);
+
+            if derive_component {
+                self.push_str("#[derive(wasmtime::component::ComponentType)]\n");
+                if lt.is_none() {
+                    self.push_str("#[derive(wasmtime::component::Lift)]\n");
+                }
+                self.push_str("#[derive(wasmtime::component::Lower)]\n");
+                self.push_str("#[component(record)]\n");
+            }
+            let mut derives = additional_derives.clone();
+            if info.is_copy() {
+                self.push_str("#[repr(C)]\n");
+                derives.extend(["Copy", "Clone"].into_iter().map(|s| s.to_string()));
+            } else if info.is_clone() {
+                derives.insert("Clone".to_string());
+            }
+            if !derives.is_empty() {
+                self.push_str("#[derive(");
+                self.push_str(&derives.into_iter().collect::<Vec<_>>().join(", "));
+                self.push_str(")]\n")
+            }
+            self.push_str(&format!("pub struct {}", name));
+            self.print_generics(lt);
+            self.push_str(" {\n");
+            for field in record.fields.iter() {
+                self.rustdoc(&field.docs);
+                if derive_component {
+                    self.push_str(&format!("#[component(name = \"{}\")]\n", field.name));
+                }
+                self.push_str("pub ");
+                self.push_str(&to_rust_ident(&field.name));
+                self.push_str(": ");
+                self.print_ty(&field.ty, mode);
+                self.push_str(",\n");
+            }
+            self.push_str("}\n");
+
+            self.push_str("impl");
+            self.print_generics(lt);
+            self.push_str(" ::core::fmt::Debug for ");
+            self.push_str(&name);
+            self.print_generics(lt);
+            self.push_str(" {\n");
+            self.push_str(
+                "fn fmt(&self, f: &mut ::core::fmt::Formatter<'_>) -> ::core::fmt::Result {\n",
+            );
+            self.push_str(&format!("f.debug_struct(\"{}\")", name));
+            for field in record.fields.iter() {
+                self.push_str(&format!(
+                    ".field(\"{}\", &self.{})",
+                    field.name,
+                    to_rust_ident(&field.name)
+                ));
+            }
+            self.push_str(".finish()\n");
+            self.push_str("}\n");
+            self.push_str("}\n");
+
+            if info.error {
+                self.push_str("impl");
+                self.print_generics(lt);
+                self.push_str(" ::core::fmt::Display for ");
+                self.push_str(&name);
+                self.print_generics(lt);
+                self.push_str(" {\n");
+                self.push_str(
+                    "fn fmt(&self, f: &mut ::core::fmt::Formatter<'_>) -> ::core::fmt::Result {\n",
+                );
+                self.push_str("write!(f, \"{:?}\", self)\n");
+                self.push_str("}\n");
+                self.push_str("}\n");
+                if self.gen.opts.std_feature {
+                    self.push_str("#[cfg(feature = \"std\")]");
+                }
+                self.push_str("impl std::error::Error for ");
+                self.push_str(&name);
+                self.push_str("{}\n");
+            }
+        }
+    }
+
+    fn print_typedef_tuple(&mut self, id: TypeId, tuple: &Tuple, docs: &Docs) {
+        let info = self.info(id);
+        for (name, mode) in self.modes_of(id) {
+            let lt = self.lifetime_for(&info, mode);
+            self.rustdoc(docs);
+            self.push_str(&format!("pub type {}", name));
+            self.print_generics(lt);
+            self.push_str(" = (");
+            for ty in tuple.types.iter() {
+                self.print_ty(ty, mode);
+                self.push_str(",");
+            }
+            self.push_str(");\n");
+        }
+    }
+
+    fn print_typedef_variant(
+        &mut self,
+        id: TypeId,
+        variant: &Variant,
+        docs: &Docs,
+        derive_component: bool,
+    ) where
+        Self: Sized,
+    {
+        self.print_rust_enum(
+            id,
+            variant.cases.iter().map(|c| {
+                (
+                    c.name.to_upper_camel_case(),
+                    Some(c.name.clone()),
+                    &c.docs,
+                    c.ty.as_ref(),
+                )
+            }),
+            docs,
+            if derive_component {
+                Some("variant")
+            } else {
+                None
+            },
+        );
+    }
+
+    fn print_rust_enum<'b>(
+        &mut self,
+        id: TypeId,
+        cases: impl IntoIterator<Item = (String, Option<String>, &'b Docs, Option<&'b Type>)> + Clone,
+        docs: &Docs,
+        derive_component: Option<&str>,
+    ) where
+        Self: Sized,
+    {
+        let info = self.info(id);
+        // We use a BTree set to make sure we don't have any duplicates and have a stable order
+        let additional_derives: BTreeSet<String> = self
+            .gen
+            .opts
+            .additional_derive_attributes
+            .iter()
+            .cloned()
+            .collect();
+        for (name, mode) in self.modes_of(id) {
+            self.rustdoc(docs);
+            let lt = self.lifetime_for(&info, mode);
+            if let Some(derive_component) = derive_component {
+                self.push_str("#[derive(wasmtime::component::ComponentType)]\n");
+                if lt.is_none() {
+                    self.push_str("#[derive(wasmtime::component::Lift)]\n");
+                }
+                self.push_str("#[derive(wasmtime::component::Lower)]\n");
+                self.push_str(&format!("#[component({})]\n", derive_component));
+            }
+            let mut derives = additional_derives.clone();
+            if info.is_copy() {
+                derives.extend(["Copy", "Clone"].into_iter().map(|s| s.to_string()));
+            } else if info.is_clone() {
+                derives.insert("Clone".to_string());
+            }
+            if !derives.is_empty() {
+                self.push_str("#[derive(");
+                self.push_str(&derives.into_iter().collect::<Vec<_>>().join(", "));
+                self.push_str(")]\n")
+            }
+            self.push_str(&format!("pub enum {name}"));
+            self.print_generics(lt);
+            self.push_str("{\n");
+            for (case_name, component_name, docs, payload) in cases.clone() {
+                self.rustdoc(docs);
+                if derive_component.is_some() {
+                    if let Some(n) = component_name {
+                        self.push_str(&format!("#[component(name = \"{}\")] ", n));
+                    }
+                }
+                self.push_str(&case_name);
+                if let Some(ty) = payload {
+                    self.push_str("(");
+                    self.print_ty(ty, mode);
+                    self.push_str(")")
+                }
+                self.push_str(",\n");
+            }
+            self.push_str("}\n");
+
+            self.print_rust_enum_debug(
+                id,
+                mode,
+                &name,
+                cases
+                    .clone()
+                    .into_iter()
+                    .map(|(name, _attr, _docs, ty)| (name, ty)),
+            );
+
+            if info.error {
+                self.push_str("impl");
+                self.print_generics(lt);
+                self.push_str(" ::core::fmt::Display for ");
+                self.push_str(&name);
+                self.print_generics(lt);
+                self.push_str(" {\n");
+                self.push_str(
+                    "fn fmt(&self, f: &mut ::core::fmt::Formatter<'_>) -> ::core::fmt::Result {\n",
+                );
+                self.push_str("write!(f, \"{:?}\", self)\n");
+                self.push_str("}\n");
+                self.push_str("}\n");
+                self.push_str("\n");
+
+                if self.gen.opts.std_feature {
+                    self.push_str("#[cfg(feature = \"std\")]");
+                }
+                self.push_str("impl");
+                self.print_generics(lt);
+                self.push_str(" std::error::Error for ");
+                self.push_str(&name);
+                self.print_generics(lt);
+                self.push_str(" {}\n");
+            }
+        }
+    }
+
+    fn print_rust_enum_debug<'b>(
+        &mut self,
+        id: TypeId,
+        mode: TypeMode,
+        name: &str,
+        cases: impl IntoIterator<Item = (String, Option<&'b Type>)>,
+    ) where
+        Self: Sized,
+    {
+        let info = self.info(id);
+        let lt = self.lifetime_for(&info, mode);
+        self.push_str("impl");
+        self.print_generics(lt);
+        self.push_str(" ::core::fmt::Debug for ");
+        self.push_str(name);
+        self.print_generics(lt);
+        self.push_str(" {\n");
+        self.push_str(
+            "fn fmt(&self, f: &mut ::core::fmt::Formatter<'_>) -> ::core::fmt::Result {\n",
+        );
+        self.push_str("match self {\n");
+        for (case_name, payload) in cases {
+            self.push_str(name);
+            self.push_str("::");
+            self.push_str(&case_name);
+            if payload.is_some() {
+                self.push_str("(e)");
+            }
+            self.push_str(" => {\n");
+            self.push_str(&format!("f.debug_tuple(\"{}::{}\")", name, case_name));
+            if payload.is_some() {
+                self.push_str(".field(e)");
+            }
+            self.push_str(".finish()\n");
+            self.push_str("}\n");
+        }
+        self.push_str("}\n");
+        self.push_str("}\n");
+        self.push_str("}\n");
+    }
+
+    fn print_typedef_option(&mut self, id: TypeId, payload: &Type, docs: &Docs) {
+        let info = self.info(id);
+
+        for (name, mode) in self.modes_of(id) {
+            self.rustdoc(docs);
+            let lt = self.lifetime_for(&info, mode);
+            self.push_str(&format!("pub type {}", name));
+            self.print_generics(lt);
+            self.push_str("= Option<");
+            self.print_ty(payload, mode);
+            self.push_str(">;\n");
+        }
+    }
+
+    fn print_typedef_result(&mut self, id: TypeId, result: &Result_, docs: &Docs) {
+        let info = self.info(id);
+
+        for (name, mode) in self.modes_of(id) {
+            self.rustdoc(docs);
+            let lt = self.lifetime_for(&info, mode);
+            self.push_str(&format!("pub type {}", name));
+            self.print_generics(lt);
+            self.push_str("= Result<");
+            self.print_optional_ty(result.ok.as_ref(), mode);
+            self.push_str(",");
+            self.print_optional_ty(result.err.as_ref(), mode);
+            self.push_str(">;\n");
+        }
+    }
+
+    fn print_typedef_enum(
+        &mut self,
+        id: TypeId,
+        name: &str,
+        enum_: &Enum,
+        docs: &Docs,
+        attrs: &[String],
+        case_attr: Box<dyn Fn(&EnumCase) -> String>,
+    ) where
+        Self: Sized,
+    {
+        let info = self.info(id);
+
+        let name = to_upper_camel_case(name);
+        self.rustdoc(docs);
+        for attr in attrs {
+            self.push_str(&format!("{}\n", attr));
+        }
+        self.push_str("#[repr(");
+        self.int_repr(enum_.tag());
+        self.push_str(")]\n");
+        // We use a BTree set to make sure we don't have any duplicates and a stable order
+        let mut derives: BTreeSet<String> = self
+            .gen
+            .opts
+            .additional_derive_attributes
+            .iter()
+            .cloned()
+            .collect();
+        derives.extend(
+            ["Clone", "Copy", "PartialEq", "Eq"]
+                .into_iter()
+                .map(|s| s.to_string()),
+        );
+        self.push_str("#[derive(");
+        self.push_str(&derives.into_iter().collect::<Vec<_>>().join(", "));
+        self.push_str(")]\n");
+        self.push_str(&format!("pub enum {name} {{\n"));
+        for case in enum_.cases.iter() {
+            self.rustdoc(&case.docs);
+            self.push_str(&case_attr(case));
+            self.push_str(&case.name.to_upper_camel_case());
+            self.push_str(",\n");
+        }
+        self.push_str("}\n");
+
+        // Auto-synthesize an implementation of the standard `Error` trait for
+        // error-looking types based on their name.
+        if info.error {
+            self.push_str("impl ");
+            self.push_str(&name);
+            self.push_str("{\n");
+
+            self.push_str("pub fn name(&self) -> &'static str {\n");
+            self.push_str("match self {\n");
+            for case in enum_.cases.iter() {
+                self.push_str(&name);
+                self.push_str("::");
+                self.push_str(&case.name.to_upper_camel_case());
+                self.push_str(" => \"");
+                self.push_str(case.name.as_str());
+                self.push_str("\",\n");
+            }
+            self.push_str("}\n");
+            self.push_str("}\n");
+
+            self.push_str("pub fn message(&self) -> &'static str {\n");
+            self.push_str("match self {\n");
+            for case in enum_.cases.iter() {
+                self.push_str(&name);
+                self.push_str("::");
+                self.push_str(&case.name.to_upper_camel_case());
+                self.push_str(" => \"");
+                if let Some(contents) = &case.docs.contents {
+                    self.push_str(contents.trim());
+                }
+                self.push_str("\",\n");
+            }
+            self.push_str("}\n");
+            self.push_str("}\n");
+
+            self.push_str("}\n");
+
+            self.push_str("impl ::core::fmt::Debug for ");
+            self.push_str(&name);
+            self.push_str(
+                "{\nfn fmt(&self, f: &mut ::core::fmt::Formatter<'_>) -> ::core::fmt::Result {\n",
+            );
+            self.push_str("f.debug_struct(\"");
+            self.push_str(&name);
+            self.push_str("\")\n");
+            self.push_str(".field(\"code\", &(*self as i32))\n");
+            self.push_str(".field(\"name\", &self.name())\n");
+            self.push_str(".field(\"message\", &self.message())\n");
+            self.push_str(".finish()\n");
+            self.push_str("}\n");
+            self.push_str("}\n");
+
+            self.push_str("impl ::core::fmt::Display for ");
+            self.push_str(&name);
+            self.push_str(
+                "{\nfn fmt(&self, f: &mut ::core::fmt::Formatter<'_>) -> ::core::fmt::Result {\n",
+            );
+            self.push_str("write!(f, \"{} (error {})\", self.name(), *self as i32)\n");
+            self.push_str("}\n");
+            self.push_str("}\n");
+            self.push_str("\n");
+            if self.gen.opts.std_feature {
+                self.push_str("#[cfg(feature = \"std\")]");
+            }
+            self.push_str("impl std::error::Error for ");
+            self.push_str(&name);
+            self.push_str("{}\n");
+        } else {
+            self.print_rust_enum_debug(
+                id,
+                TypeMode::Owned,
+                &name,
+                enum_
+                    .cases
+                    .iter()
+                    .map(|c| (c.name.to_upper_camel_case(), None)),
+            )
+        }
+    }
+
+    fn print_typedef_alias(&mut self, id: TypeId, ty: &Type, docs: &Docs) {
+        if self.is_exported_resource(id) {
+            let target = dealias(self.resolve, id);
+            let ty = &self.resolve.types[target];
+            // TODO: We could wait until we know how a resource (and its
+            // aliases) is used prior to generating declarations.  For example,
+            // if only borrows are used, no need to generate the `Own{name}`
+            // version.
+            self.mark_resource_owned(target);
+            for prefix in ["Own", ""] {
+                self.rustdoc(docs);
+                self.push_str(&format!(
+                    "pub type {prefix}{} = {};\n",
+                    self.resolve.types[id]
+                        .name
+                        .as_deref()
+                        .unwrap()
+                        .to_upper_camel_case(),
+                    self.type_path_with_name(
+                        target,
+                        format!(
+                            "{prefix}{}",
+                            ty.name.as_deref().unwrap().to_upper_camel_case()
+                        )
+                    )
+                ));
+            }
+        } else {
+            let info = self.info(id);
+            for (name, mode) in self.modes_of(id) {
+                self.rustdoc(docs);
+                self.push_str(&format!("pub type {name}"));
+                let lt = self.lifetime_for(&info, mode);
+                self.print_generics(lt);
+                self.push_str(" = ");
+                self.print_ty(ty, mode);
+                self.push_str(";\n");
+            }
+        }
+    }
+
+    fn print_type_list(&mut self, id: TypeId, ty: &Type, docs: &Docs) {
+        let info = self.info(id);
+        for (name, mode) in self.modes_of(id) {
+            let lt = self.lifetime_for(&info, mode);
+            self.rustdoc(docs);
+            self.push_str(&format!("pub type {}", name));
+            self.print_generics(lt);
+            self.push_str(" = ");
+            self.print_list(ty, mode);
+            self.push_str(";\n");
+        }
+    }
+
+    fn param_name(&self, ty: TypeId) -> String {
+        let info = self.info(ty);
+        let name = to_upper_camel_case(self.resolve.types[ty].name.as_ref().unwrap());
+        if self.uses_two_names(&info) {
+            format!("{}Param", name)
+        } else {
+            name
+        }
+    }
+
+    fn result_name(&self, ty: TypeId) -> String {
+        let info = self.info(ty);
+        let name = to_upper_camel_case(self.resolve.types[ty].name.as_ref().unwrap());
+        if self.uses_two_names(&info) {
+            format!("{}Result", name)
+        } else if self.is_exported_resource(ty) {
+            format!("Own{name}")
+        } else {
+            name
+        }
+    }
+
+    fn uses_two_names(&self, info: &TypeInfo) -> bool {
+        // Types are only duplicated if explicitly requested ...
+        matches!(
+            self.gen.opts.ownership,
+            Ownership::Borrowing {
+                duplicate_if_necessary: true
+            }
+        )
+            // ... and if they're both used in a borrowed/owned context
+            && info.borrowed
+            && info.owned
+            // ... and they have a list ...
+            && info.has_list
+            // ... and if there's NOT an `own` handle since those are always
+            // done by ownership.
+            && !info.has_own_handle
+    }
+
+    fn lifetime_for(&self, info: &TypeInfo, mode: TypeMode) -> Option<&'static str> {
+        let lt = match mode {
+            TypeMode::AllBorrowed(s) | TypeMode::LeafBorrowed(s) | TypeMode::HandlesBorrowed(s) => {
+                s
+            }
+            TypeMode::Owned => return None,
+        };
+        if info.has_borrow_handle {
+            return Some(lt);
+        }
+        if matches!(self.gen.opts.ownership, Ownership::Owning) {
+            return None;
+        }
+        // No lifetimes needed unless this has a list.
+        if !info.has_list {
+            return None;
+        }
+        // If two names are used then this type will have an owned and a
+        // borrowed copy and the borrowed copy is being used, so it needs a
+        // lifetime. Otherwise if it's only borrowed and not owned then this can
+        // also use a lifetime since it's not needed in two contexts and only
+        // the borrowed version of the structure was generated.
+        if self.uses_two_names(info) || (info.borrowed && !info.owned) {
+            Some(lt)
+        } else {
+            None
+        }
+    }
+
+    // fn ownership(&self) -> Ownership {
+    //     self.gen.opts.ownership
+    // }
 
     fn path_to_interface(&self, interface: InterfaceId) -> Option<String> {
         let mut path = String::new();
@@ -497,14 +1506,6 @@ impl<'a> RustGenerator<'a> for InterfaceGenerator<'a> {
         let name = &self.gen.interface_names[&interface];
         path.push_str(name);
         Some(path)
-    }
-
-    fn std_feature(&self) -> bool {
-        self.gen.opts.std_feature
-    }
-
-    fn use_raw_strings(&self) -> bool {
-        self.gen.opts.raw_strings
     }
 
     fn push_vec_name(&mut self) {
@@ -530,7 +1531,7 @@ impl<'a> RustGenerator<'a> for InterfaceGenerator<'a> {
         )
     }
 
-    fn mark_resource_owned(&mut self, resource: TypeId) {
+    pub fn mark_resource_owned(&mut self, resource: TypeId) {
         self.gen
             .resources
             .entry(dealias(self.resolve, resource))
@@ -551,10 +1552,6 @@ impl<'a> RustGenerator<'a> for InterfaceGenerator<'a> {
 
     fn info(&self, ty: TypeId) -> TypeInfo {
         self.gen.types.get(ty)
-    }
-
-    fn types_mut(&mut self) -> &mut Types {
-        &mut self.gen.types
     }
 
     fn print_borrowed_slice(
